# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My User Project CRUD",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "165.232.172.82:443",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],

  "paths": {
    "/users/all": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users from database.",
        "responses": {
          "200": {
            "description": "Return an array of all Users",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        }
      }
    },
    "/users/{userID}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by it's ID from database",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "type": "integer",
            "description": "The ID of the user that we want to get",
            "schema": {
              "$ref": "#/definitions/userID"
            }
          }
        ],
        "produces": ["application/json", "text/plain"],
        "responses": {
          "200": {
            "description": "Return the user with matching ID from database.",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "204": {
            "description": "No user with id:[getById] found in database."
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user in database",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "User that we want to create. ***NOTE*** Don't include the ID as the ID feild will be automatically generated by our API",
            "schema": {
              "$ref": "#/definitions/newUser"
            }
          }
        ],
        "produces": ["application/json", "text/plain"],
        "responses": {
          "200": {
            "description": "Return the new user with the ID field and it's value.",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "204": {
            "description": "Internal error, No new user created."
          }
        }
      }
    },
    "/users/{userID}/{newJobTitle}": {
      "patch": {
        "tags": ["Users"],
        "summary": "Update user's job title from it's ID",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "type": "integer",
            "description": "The ID of the user that we want to get",
            "schema": {
              "$ref": "#/definitions/userID"
            }
          },
          {
            "name": "newJobTitle",
            "in": "path",
            "type": "string",
            "description": "The new job title for the user that we want to update",
            "schema": {
              "$ref": "#/definitions/userID"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Return updated user from database"
          },
          "204": {
            "description": "No user with id:[getById] found in database."
          }
        }
      }
    },
    "/users/{deleteById}": {
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by ID from database",
        "parameters": [
          {
            "name": "deleteById",
            "in": "path",
            "description": "User ID that you want to delete",
            "schema": {
              "$ref": "#/definitions/userID"
            }
          }
        ],
        "produces": ["text/plain"],
        "responses": {
          "200": {
            "description": "User with id:[deleteById] deleted from database"
          },
          "204": {
            "description": "No user with id:[deleteById] found in database."
          }
        }
      }
    }
  },
  "definitions": {
    "user": {
      "required": ["id", "firstName", "lastName", "jobTitle"],
      "properties": {
        "id": {
          "type": "integer",
          "uniqueItems": true
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "jobTitle": {
          "type": "string"
        }
      }
    },
    "newUser": {
      "required": ["firstName", "lastName", "jobTitle"],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "jobTitle": {
          "type": "string"
        }
      }
    },
    "userID": {
      "required": ["id"],
      "properties": {
        "id": {
          "type": "integer",
          "uniqueItems": true
        }
      }
    }
  }
}